requirements.txt
asgiref==3.2.10
Django==3.1
django-rest-auth==0.9.5
djangorestframework==3.11.1
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
